MCGA-145: Major refactor of navigationfragment. Add GPSManager
