+++ app/src/main/res/layout/nav_main_fragment.xml
@@ -10,10 +10,6 @@
     tools:layout="@layout/activity_main"
     android:focusableInTouchMode="true">
 
-    <include
-        android:id="@+id/nav_toolbar"
-        layout="@layout/nav_main_toolbar"/>
-
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
